fix(ci): 修复 tenant_id 类型不匹配（i32 → i64）
CI #1424 构建失败：extract_tenant_id 返回 i32，但 product_color_price
模型的 tenant_id 字段为 i64。添加 as i64 类型转换。

+++ backend/src/handlers/quotation_handler.rs
@@ -456,7 +456,7 @@ pub async fn set_color_price(
     use sea_orm::ActiveModelTrait;
     // BE-F-1/BE-C-7 修复（2026-06-25 第二次全面审计）：
     // 原 tenant_id 硬编码为 1，违反租户隔离规范。改用 extract_tenant_id 从认证上下文提取。
-    let tenant_id = extract_tenant_id(&auth)?;
+    let tenant_id = extract_tenant_id(&auth)? as i64;
     let mut active: product_color_price::ActiveModel = product_color_price::Model {
         id: payload.id.unwrap_or(0),
         product_id: payload.product_id,
